+++ src/components/settings/SettingsReturnHomeField.jsx
@@ -5,7 +5,38 @@ import BackIcon from '@material-symbols/svg-400/rounded/arrow_back.svg';
 import { getDocumentSettingsById, useDocumentStore } from '@/stores/document';
 import { useCurrentDocumentId, useSetUserCursor } from '@/stores/user';
 
+import SettingsFieldButton from './SettingsFieldButton';
+
 export default function SettingsReturnHomeField() {
+  const navigate = useNavigate();
+  const documentId = useCurrentDocumentId();
+  const projectId = useDocumentStore(
+    (ctx) => getDocumentSettingsById(ctx, documentId)?.projectId,
+  );
+  const setUserCursor = useSetUserCursor();
+  const deleteDocument = useDocumentStore((ctx) => ctx.deleteDocument);
+
+  function onClick() {
+    setUserCursor('', '', '');
+    if (!projectId) {
+      // Since no project id exists yet, just delete this project.
+      deleteDocument(documentId);
+    }
+    navigate('/');
+  }
+
+  return (
+    <SettingsFieldButton
+      className="outline-none -mt-3"
+      Icon={BackIcon}
+      onClick={onClick}>
+      Exit to home
+    </SettingsFieldButton>
+  );
+}
+
+// @ts-expect-error Not used, but it is nice :3
+function SettingsProjectBackButton() {
   const navigate = useNavigate();
   const documentId = useCurrentDocumentId();
   const projectId = useDocumentStore(