+++ src/components/takes/TakeEntryNew.jsx
@@ -1,39 +1,30 @@
 import TakeEntryDetails from './TakeEntryDetails';
-import TakeEntryHeader from './TakeEntryHeader';
+import TakeEntryHeaderNew from './TakeEntryHeaderNew';
 import TakeEntryLayout from './TakeEntryLayout';
 
 /**
  * @param {object} props
- * @param {string} [props.className]
  * @param {import('@/stores/document/DocumentStore').DocumentId} props.documentId
  * @param {import('@/stores/document/DocumentStore').SceneId} props.sceneId
  * @param {import('@/stores/document/DocumentStore').ShotId} props.shotId
- * @param {import('@/stores/document/DocumentStore').TakeId} props.takeId
- * @param {boolean} props.cloudExportable
  * @param {'list'|'inline'} props.viewMode
  */
-export default function TakeEntry({
-  className,
+export default function TakeEntryNew({
   documentId,
   sceneId,
   shotId,
-  takeId,
   viewMode,
 }) {
   return (
-    <TakeEntryLayout className={className} viewMode={viewMode}>
-      <TakeEntryHeader
+    <TakeEntryLayout viewMode={viewMode}>
+      <TakeEntryHeaderNew
+        className="text-gray-400"
         documentId={documentId}
         sceneId={sceneId}
         shotId={shotId}
-        takeId={takeId}
         viewMode={viewMode}
       />
-      <TakeEntryDetails
-        documentId={documentId}
-        takeId={takeId}
-        viewMode={viewMode}
-      />
+      <TakeEntryDetails documentId={documentId} takeId="" viewMode={viewMode} />
     </TakeEntryLayout>
   );
 }