+++ src/libs/ImageWithCaption.jsx
@@ -2,7 +2,7 @@
  * @param {object} props
  * @param {string} props.src
  * @param {string} props.alt
- * @param {import('react').FC<any>} props.Icon
+ * @param {import('react').FC<any>|null} props.Icon
  * @param {string} [props.caption]
  * @param {string} [props.className]
  * @param {import('react').ReactNode} [props.children]
@@ -21,7 +21,8 @@ export default function ImageWithCaption({
         'relative flex items-center overflow-hidden whitespace-nowrap' +
         ' ' +
         className
-      }>
+      }
+      title={alt}>
       {src ? (
         <img
           className="flex-1 object-contain"
@@ -29,9 +30,9 @@ export default function ImageWithCaption({
           alt={alt}
           loading="lazy"
         />
-      ) : (
+      ) : Icon ? (
         <Icon className="flex-1 fill-gray-400" />
-      )}
+      ) : null}
       {caption && (
         <figcaption className="absolute right-2 bottom-0 text-right font-mono [text-shadow:_-1px_-1px_2px_white,_-1px_1px_2px_white,_1px_1px_2px_white,_1px_-1px_2px_white]">
           {caption}