+++ CHANGELOG.md
@@ -3,10 +3,25 @@
 - Photo recorder for story building
 - Shot in text.
 - Zoom requires it ZOOMS out before zooming back in?
+- Record Take 4
+- shot details focus experience
+- - nav takes on shotlist
+- - nav shots on edit mode
+- move the exit button
+- move it higher avoid bar
+- add fav last take Take 3 (-> 4)
+- Good take on filenames?
+- - Searchable
+- Batch export
 
 - Add shot number for both sides
 
 - Change to full-width storyboard on smaller screens
+- Change `Record!` to include next take number
+- Change take rating button to include take number
+- Change take rating to only be good or not
+- Change record button to placeholder take
+- Change exit button to settings
 
 - Fix overflow for scene headers